+++ extensions/finnish-dictionary/src/general-lookup.tsx
@@ -1,0 +1,93 @@
+import { useEffect, useState } from "react";
+import { Action, ActionPanel, List } from "@raycast/api";
+import WordDictionary from "./word-dictionary";
+
+const languages: Record<string, string> = {
+  Hungarian: "hun",
+  Swedish: "swe",
+  Icelandic: "isl",
+  Hebrew: "heb",
+  Chinese: "cmn",
+  Estonian: "est",
+  Finnish: "fin",
+  Hindi: "hin",
+  Polish: "pol",
+  Korean: "kor",
+  Turkish: "tur",
+  Czech: "ces",
+  Romanian: "ron",
+  Danish: "dan",
+  Norwegian: "nor",
+  Russian: "rus",
+  Arabic: "ara",
+  Portuguese: "por",
+  Vietnamese: "vie",
+  French: "fra",
+  German: "deu",
+  Lithuanian: "lit",
+  Italian: "ita",
+  Persian: "fas",
+  "Min Nan": "nan",
+  Ukrainian: "ukr",
+  Indonesian: "ind",
+  English: "eng",
+  Latvian: "lav",
+  Greek: "ell",
+  Spanish: "spa",
+  Urdu: "urd",
+  Dutch: "nld",
+  Japanese: "jpn",
+  Afrikaans: "afr",
+  Thai: "tha",
+  Croatian: "hrv",
+  Serbian: "srp",
+  Slovak: "slk",
+  Bulgarian: "bul",
+  Malay: "zlm",
+  Slovene: "slv",
+};
+
+export default function Command() {
+  const [searchText, setSearchText] = useState("");
+  const [langs, setLangs] = useState("");
+  const [filteredList, filterList] = useState<string[]>([]);
+
+  useEffect(() => {
+    filterList(Object.keys(languages).filter((item) => item.toLowerCase().includes(searchText.toLowerCase())));
+  }, [searchText]);
+
+  const selectLanguage = (lang: string) => {
+    lang = languages[lang] as string;
+    setSearchText("");
+    setTimeout(() => {
+      if (langs.length === 0) {
+        setLangs(lang);
+      } else {
+        setLangs(langs + "-" + lang);
+      }
+    }, 10);
+  };
+
+  return langs.indexOf("-") === -1 ? (
+    <List
+      searchText={searchText}
+      onSearchTextChange={setSearchText}
+      navigationTitle={`Define ${langs} word`}
+      searchBarPlaceholder="Select languages"
+    >
+      {filteredList.map((item) => (
+        <List.Item
+          key={item}
+          title={item}
+          actions={
+            <ActionPanel>
+              <Action title="Select" onAction={() => selectLanguage(item)} />
+            </ActionPanel>
+          }
+        />
+      ))}
+    </List>
+  ) : (
+    <WordDictionary from={langs.split("-")[0]} to={langs.split("-")[1]} />
+  );
+}