+++ extensions/finnish-dictionary/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Finnish Dictionary Changelog
 
+
+## [Added any language capability] - 2022-06-03
+
+* Add a new command to search terms in other languages
+
 ## [Added History capability] - 2022-05-25
 
 * Add caching of previous searches for quick lookups